+++ datafiles/tests/test_managers.py
@@ -1,16 +1,17 @@
-# pylint: disable=unused-variable,protected-access
+# pylint: disable=unused-variable
 
 from dataclasses import dataclass
 from pathlib import Path
+from unittest.mock import patch
 
 import pytest
 
 from datafiles import managers
-from datafiles.models import ModelMeta
+from datafiles.models import ModelMeta, create_model
 
 
 @dataclass
-class MyModel:
+class MyClass:
     foobar: int
 
 
@@ -20,11 +21,39 @@ class MyField:
         return python_value
 
 
+def describe_model_manager():
+    @pytest.fixture
+    def manager():
+        return managers.Manager(create_model(MyClass, manual=True))
+
+    def describe_get_or_none():
+        @patch('datafiles.managers.Datafile.exists', True)
+        def when_file_exists(expect, manager):
+            expect(manager.get_or_none(foobar=1)) == MyClass(foobar=1)
+
+        @patch('datafiles.managers.Datafile.exists', False)
+        def when_file_missing(expect, manager):
+            expect(manager.get_or_none(foobar=2)) == None
+
+    def describe_get_or_create():
+        @patch('datafiles.managers.Datafile.save')
+        @patch('datafiles.managers.Datafile.exists', True)
+        def when_file_exists(mock_save, expect, manager):
+            expect(manager.get_or_create(foobar=1)) == MyClass(foobar=1)
+            expect(mock_save.call_count) == 0
+
+        @patch('datafiles.managers.Datafile.save')
+        @patch('datafiles.managers.Datafile.exists', False)
+        def when_file_missing(mock_save, expect, manager):
+            expect(manager.get_or_create(foobar=2)) == MyClass(foobar=2)
+            expect(mock_save.call_count) == 1
+
+
 def describe_instance_manager():
     @pytest.fixture
     def manager():
         return managers.Datafile(
-            instance=MyModel(foobar=42),
+            instance=MyClass(foobar=42),
             attrs={},
             pattern=None,
             manual=ModelMeta.datafile_manual,